Bug bash - Refactor RBM-related utils
Merge numpy_splitters into python_splitters
Refactor AffinityMatrix class
