+++ tests/unit/test_python_splitter.py
@@ -15,6 +15,7 @@ from reco_utils.dataset.python_splitters import (
     python_chrono_split,
     python_random_split,
     python_stratified_split,
+    numpy_stratified_split,
 )
 
 from reco_utils.common.constants import (
@@ -35,6 +36,9 @@ def test_specs():
         "ratios": [0.2, 0.3, 0.5],
         "split_numbers": [2, 3, 5],
         "tolerance": 0.01,
+        "number_of_items": 50,
+        "number_of_users": 20,
+        "fluctuation": 0.02,
     }
 
 
@@ -253,3 +257,116 @@ def test_stratified_splitter(test_specs, python_dataset):
         test_specs["ratios"][2], test_specs["tolerance"]
     )
 
+
+@pytest.fixture(scope="module")
+def python_int_dataset(test_specs):
+    # fix the the random seed
+    np.random.seed(test_specs["seed"])
+
+    # generates the user/item affinity matrix. Ratings are from 1 to 5, with 0s denoting unrated items
+    return np.random.randint(
+        low=0,
+        high=6,
+        size=(test_specs["number_of_users"], test_specs["number_of_items"]),
+    )
+
+
+@pytest.fixture(scope="module")
+def python_float_dataset(test_specs):
+    # fix the the random seed
+    np.random.seed(test_specs["seed"])
+
+    # generates the user/item affinity matrix. Ratings are from 1 to 5, with 0s denoting unrated items
+    return np.random.random(
+            size=(test_specs["number_of_users"], test_specs["number_of_items"])
+        ) * 5
+
+
+def test_int_numpy_stratified_splitter(test_specs, python_int_dataset):
+    # generate a syntetic dataset
+    X = python_int_dataset
+
+    # the splitter returns (in order): train and test user/affinity matrices, train and test datafarmes and user/items to matrix maps
+    Xtr, Xtst = numpy_stratified_split(
+        X, ratio=test_specs["ratio"], seed=test_specs["seed"]
+    )
+
+    # check that the generated matrices have the correct dimensions
+    assert (Xtr.shape[0] == X.shape[0]) & (Xtr.shape[1] == X.shape[1])
+    assert (Xtst.shape[0] == X.shape[0]) & (Xtst.shape[1] == X.shape[1])
+
+    X_rated = np.sum(X != 0, axis=1)  # number of total rated items per user
+    Xtr_rated = np.sum(Xtr != 0, axis=1)  # number of rated items in the train set
+    Xtst_rated = np.sum(Xtst != 0, axis=1)  # number of rated items in the test set
+
+    # global split: check that the all dataset is split in the correct ratio
+    assert Xtr_rated.sum() / (X_rated.sum()) == pytest.approx(
+        test_specs["ratio"], test_specs["tolerance"]
+    )
+
+    assert Xtst_rated.sum() / (X_rated.sum()) == pytest.approx(
+        1 - test_specs["ratio"], test_specs["tolerance"]
+    )
+
+    # This implementation of the stratified splitter performs a random split at the single user level. Here we check
+    # that also this more stringent condition is verified. Note that user to user fluctuations in the split ratio
+    # are stronger than for the entire dataset due to the random nature of the per user splitting.
+    # For this reason we allow a slightly bigger tollerance, as specified in the test_specs()
+
+    assert (
+        (Xtr_rated / X_rated <= test_specs["ratio"] + test_specs["fluctuation"]).all()
+        & (Xtr_rated / X_rated >= test_specs["ratio"] - test_specs["fluctuation"]).all()
+    )
+
+    assert (
+        (
+            Xtst_rated / X_rated
+            <= (1 - test_specs["ratio"]) + test_specs["fluctuation"]
+        ).all()
+        & (
+            Xtst_rated / X_rated
+            >= (1 - test_specs["ratio"]) - test_specs["fluctuation"]
+        ).all()
+    )
+
+
+def test_float_numpy_stratified_splitter(test_specs, python_float_dataset):
+    # generate a syntetic dataset
+    X = python_float_dataset
+
+    # the splitter returns (in order): train and test user/affinity matrices, train and test datafarmes and user/items to matrix maps
+    Xtr, Xtst = numpy_stratified_split(
+        X, ratio=test_specs["ratio"], seed=test_specs["seed"]
+    )
+
+    # Tests
+    # check that the generated matrices have the correct dimensions
+    assert (Xtr.shape[0] == X.shape[0]) & (Xtr.shape[1] == X.shape[1])
+
+    assert (Xtst.shape[0] == X.shape[0]) & (Xtst.shape[1] == X.shape[1])
+
+    X_rated = np.sum(X != 0, axis=1)  # number of total rated items per user
+    Xtr_rated = np.sum(Xtr != 0, axis=1)  # number of rated items in the train set
+    Xtst_rated = np.sum(Xtst != 0, axis=1)  # number of rated items in the test set
+
+    # global split: check that the all dataset is split in the correct ratio
+    assert Xtr_rated.sum() / (X_rated.sum()) == pytest.approx(
+        test_specs["ratio"], test_specs["tolerance"]
+    )
+
+    assert Xtst_rated.sum() / (X_rated.sum()) == pytest.approx(
+        1 - test_specs["ratio"], test_specs["tolerance"]
+    )
+
+    # This implementation of the stratified splitter performs a random split at the single user level. Here we check
+    # that also this more stringent condition is verified. Note that user to user fluctuations in the split ratio
+    # are stronger than for the entire dataset due to the random nature of the per user splitting.
+    # For this reason we allow a slightly bigger tollerance, as specified in the test_specs()
+
+    assert Xtr_rated / X_rated == pytest.approx(
+        test_specs["ratio"], rel=test_specs["fluctuation"]
+    )
+
+    assert Xtst_rated / X_rated == pytest.approx(
+        (1 - test_specs["ratio"]), rel=test_specs["fluctuation"]
+    )
